Web/EnterthePolygon/requirements.txt
Django==2.1.4
gunicorn==19.9.0
ecdsa==0.13
ExifRead==2.1.2
future==0.17.1
Pillow
pyasn1==0.4.5
pyjwt
pytz==2018.7
rsa==4.0
six==1.12.0
psycopg2
selenium
pika==0.13.1
requests
